Remove defunct test.
With the new context approach, this test is fundamentally flawed. A supervisor
takes a new subcontext on startup and is expected to pass that to its tests.
We can't assume the context is still live after a termination, but we can't
restart the services either.

So, basically, the supervisors just need to never panic. Fortunately,
knock on wood, they don't.

This closes #66.