Do not examine the returned error for the context failure. Check the context directly.
This closes Issue 77.

Issue 77 is that returning one of the distinguished context errors as a service's
error value caused the supervisor to intepret it as the context being cancelled.
This is not necessarily the case; it is easy for another context, possibly
a sub-context, to have its error bubble up to the service's return value.
Since the supervisor has its own reference to the context, it should instead
directly check if the context is canceled or not, rather than depending on the
error return from the service.